Do not auto-commit as part of the github workflow
We keep running into issues where certain blogs will change the titles
of their posts in such a way that it creates a conflict with the
previous name (due to case insensitivty at some stage of the pipeline),
and it breaks cryptic news completely. Even when the two names don't
directly conflict, we still end up with two different entries for the
same post, each with a slightly different name, and with one often being
a broken link.

I think the general solution is to stop syndicating posts like we've
been doing. Since we're just publishing straight to gh-pages anyway,
there's not much reason to commit the posts to the repo. Just
re-generate the entire content directory anew on every run of the cron,
publish that directly, and don't commit anything.

This eliminates issues with blogs changing their post titles, and makes
the commit history much cleaner.